+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapInstallmentsSelectorView.swift
@@ -7,126 +7,9 @@
 
 import UIKit
 
-typealias PXOneTapInstallmentsSelectorData = (title: NSAttributedString, value: NSAttributedString, isSelected: Bool)
-
-final class PXOneTapInstallmentsSelectorCell: UITableViewCell {
-
-    var data: PXOneTapInstallmentsSelectorData? = nil
-
-    func updateData(_ data: PXOneTapInstallmentsSelectorData) {
-        self.data = data
-        self.selectionStyle = .none
-        
-        let titleLabel = UILabel()
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel.attributedText = data.title
-        titleLabel.textAlignment = .left
-        contentView.addSubview(titleLabel)
-        PXLayout.pinLeft(view: titleLabel, withMargin: PXLayout.M_MARGIN).isActive = true
-        PXLayout.centerVertically(view: titleLabel).isActive = true
-
-        let valueLabel = UILabel()
-        valueLabel.translatesAutoresizingMaskIntoConstraints = false
-        valueLabel.attributedText = data.value
-        valueLabel.textAlignment = .right
-        contentView.addSubview(valueLabel)
-        PXLayout.pinRight(view: valueLabel, withMargin: PXLayout.M_MARGIN).isActive = true
-        PXLayout.centerVertically(view: valueLabel).isActive = true
-
-        if data.isSelected {
-            let selectedIndicatorView = UIView()
-            selectedIndicatorView.translatesAutoresizingMaskIntoConstraints = false
-            selectedIndicatorView.backgroundColor = ThemeManager.shared.getAccentColor()
-            contentView.addSubview(selectedIndicatorView)
-            PXLayout.setWidth(owner: selectedIndicatorView, width: 4).isActive = true
-            PXLayout.pinTop(view: selectedIndicatorView).isActive = true
-            PXLayout.pinBottom(view: selectedIndicatorView).isActive = true
-            PXLayout.pinLeft(view: selectedIndicatorView, withMargin: 0).isActive = true
-        }
-    }
-}
-
 protocol PXOneTapInstallmentsSelectorProtocol: NSObjectProtocol {
     func payerCostSelected(_ payerCost: PXPayerCost)
 }
-
-final class PXOneTapInstallmentsSelectorViewModel {
-    let installmentData: PXInstallment
-    let selectedPayerCost: PXPayerCost?
-
-    init(installmentData: PXInstallment, selectedPayerCost: PXPayerCost?) {
-        self.installmentData = installmentData
-        self.selectedPayerCost = selectedPayerCost
-    }
-
-    func numberOfRowsInSection(_ section: Int) -> Int {
-        return installmentData.payerCosts.count
-    }
-
-    func cellForRowAt(_ indexPath: IndexPath) -> UITableViewCell {
-        let cell = PXOneTapInstallmentsSelectorCell()
-        if let payerCost = getPayerCostForRowAt(indexPath) {
-            var isSelected = false
-            if let selectedPayerCost = selectedPayerCost, selectedPayerCost == payerCost {
-                isSelected = true
-            }
-            let data = getDataFor(payerCost: payerCost, isSelected: isSelected)
-            cell.updateData(data)
-            return cell
-        }
-        return cell
-    }
-
-    func heightForRowAt(_ indexPath: IndexPath) -> CGFloat {
-        return PXOneTapInstallmentInfoView.DEFAULT_ROW_HEIGHT
-    }
-
-    func getDataFor(payerCost: PXPayerCost, isSelected: Bool) -> PXOneTapInstallmentsSelectorData {
-        let currency = SiteManager.shared.getCurrency()
-        let showDescription = MercadoPagoCheckout.showPayerCostDescription()
-
-        var title: NSAttributedString = NSAttributedString(string: "")
-        var value: NSAttributedString = NSAttributedString(string: "")
-
-        if payerCost.installments == 1 {
-            value = NSAttributedString(string: "")
-        } else if payerCost.hasInstallmentsRate() {
-            let attributedTotal = NSMutableAttributedString(attributedString: NSAttributedString(string: "(", attributes: [NSAttributedStringKey.foregroundColor: UIColor.px_grayLight()]))
-            attributedTotal.append(Utils.getAttributedAmount(payerCost.totalAmount, currency: currency, color: UIColor.px_grayLight(), fontSize: 15, baselineOffset: 3))
-            attributedTotal.append(NSAttributedString(string: ")", attributes: [NSAttributedStringKey.foregroundColor: UIColor.px_grayLight()]))
-
-            if showDescription == false {
-                value = NSAttributedString(string: "")
-            } else {
-                value = attributedTotal
-            }
-
-        } else {
-            if showDescription == false {
-                value = NSAttributedString(string: "")
-            } else {
-                value = NSAttributedString(string: "Sin interés".localized, attributes: [NSAttributedStringKey.foregroundColor: ThemeManager.shared.noTaxAndDiscountLabelTintColor()])
-            }
-
-        }
-        var installmentNumber = String(format: "%i", payerCost.installments)
-        installmentNumber = "\(installmentNumber) x "
-        let totalAmount = Utils.getAttributedAmount(payerCost.installmentAmount, thousandSeparator: currency.getThousandsSeparatorOrDefault(), decimalSeparator: currency.getDecimalSeparatorOrDefault(), currencySymbol: currency.getCurrencySymbolOrDefault(), color: UIColor.black, centsFontSize: 14, baselineOffset: 5)
-
-        let atribute = [NSAttributedStringKey.font: Utils.getFont(size: 20), NSAttributedStringKey.foregroundColor: UIColor.black]
-        let installmentLabel = NSMutableAttributedString(string: installmentNumber, attributes: atribute)
-
-        installmentLabel.append(totalAmount)
-        title = installmentLabel
-
-        return PXOneTapInstallmentsSelectorData(title, value, isSelected)
-    }
-
-    func getPayerCostForRowAt(_ indexPath: IndexPath) -> PXPayerCost? {
-        return installmentData.payerCosts[indexPath.row]
-    }
-}
-
 
 final class PXOneTapInstallmentsSelectorView: PXComponentView, UITableViewDelegate, UITableViewDataSource {
 
@@ -152,6 +35,7 @@ final class PXOneTapInstallmentsSelectorView: PXComponentView, UITableViewDelega
 
     var tableViewHeightConstraint: NSLayoutConstraint?
     let tableViewTopSeparator = UIView()
+    let shadowView = UIImageView(image: ResourceManager.shared.getImage("one-tap-installments-shadow"))
 
     func render() {
         removeAllSubviews()
@@ -228,6 +112,20 @@ final class PXOneTapInstallmentsSelectorView: PXComponentView, UITableViewDelega
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         if let selectedPayerCost = model.getPayerCostForRowAt(indexPath) {
             delegate?.payerCostSelected(selectedPayerCost)
+        }
+    }
+
+    func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        shadowView.translatesAutoresizingMaskIntoConstraints = false
+        if tableView.contentOffset.y > CGFloat(0.0) {
+            self.addSubview(shadowView)
+            PXLayout.pinRight(view: shadowView).isActive = true
+            PXLayout.pinLeft(view: shadowView).isActive = true
+            PXLayout.pinTop(view: shadowView).isActive = true
+            PXLayout.setHeight(owner: shadowView, height: 22).isActive = true
+        } else {
+            shadowView.removeFromSuperview()
+            shadowView.removeConstraints(shadowView.constraints)
         }
     }
 }
