+++ MercadoPagoSDK/MercadoPagoSDK/RevampImages.xcassets/PXDefaultIcons/one-tap-installments-shadow.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "one-tap-installments-shadow.pdf"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
